Move tests to pagination module
